requirements.txt
requests>=2.28.0
tldextract>=3.4.0
python-whois>=0.9.0
pytest>=7.0.0
